Unimportant modifications
Changed the version number of the old version to make it more standardized, to 'Am.n.s_x.y', where A is the Alpha version, m is the major version, n is the minor version, s is the backup version, x.y refers to the update in x months, and the y update in total. Added 'Changelog' folder.